Move parser classes to separate package
