+++ modules/compiler/src/main/java/config/parser/ConfigParserPluginFactory.java
@@ -13,15 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nextflow.script.ast;
+package nextflow.config.parser;
 
 import org.codehaus.groovy.control.ParserPlugin;
 import org.codehaus.groovy.control.ParserPluginFactory;
 
-public class ScriptParserPluginFactory extends ParserPluginFactory {
+public class ConfigParserPluginFactory extends ParserPluginFactory {
 
     @Override
     public ParserPlugin createParserPlugin() {
-        return new ScriptParserPlugin();
+        return new ConfigParserPlugin();
     }
 }
