+++ modules/compiler/src/main/java/script/parser/ScriptParserPluginFactory.java
@@ -13,15 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nextflow.config.ast;
+package nextflow.script.parser;
 
 import org.codehaus.groovy.control.ParserPlugin;
 import org.codehaus.groovy.control.ParserPluginFactory;
 
-public class ConfigParserPluginFactory extends ParserPluginFactory {
+public class ScriptParserPluginFactory extends ParserPluginFactory {
 
     @Override
     public ParserPlugin createParserPlugin() {
-        return new ConfigParserPlugin();
+        return new ScriptParserPlugin();
     }
 }
