+++ modules/language-server/src/main/java/nextflow/lsp/services/config/ConfigAstCache.java
@@ -26,8 +26,8 @@ import nextflow.config.ast.ConfigBlockNode;
 import nextflow.config.ast.ConfigIncludeNode;
 import nextflow.config.ast.ConfigIncompleteNode;
 import nextflow.config.ast.ConfigNode;
-import nextflow.config.ast.ConfigParserPluginFactory;
 import nextflow.config.ast.ConfigVisitorSupport;
+import nextflow.config.parser.ConfigParserPluginFactory;
 import nextflow.lsp.ast.ASTNodeCache;
 import nextflow.lsp.ast.ASTParentVisitor;
 import nextflow.lsp.compiler.Compiler;
