+++ modules/language-server/src/main/java/nextflow/lsp/services/script/ScriptAstCache.java
@@ -44,13 +44,13 @@ import nextflow.script.ast.OutputNode;
 import nextflow.script.ast.ParamNode;
 import nextflow.script.ast.ProcessNode;
 import nextflow.script.ast.ScriptNode;
-import nextflow.script.ast.ScriptParserPluginFactory;
 import nextflow.script.ast.ScriptVisitorSupport;
 import nextflow.script.ast.WorkflowNode;
 import nextflow.script.control.PhaseAware;
 import nextflow.script.control.Phases;
 import nextflow.script.control.ResolveVisitor;
 import nextflow.script.control.VariableScopeVisitor;
+import nextflow.script.parser.ScriptParserPluginFactory;
 import org.codehaus.groovy.GroovyBugError;
 import org.codehaus.groovy.ast.ASTNode;
 import org.codehaus.groovy.ast.ClassNode;
